requirements.txt
pydantic==2.5.0
pymongo==4.6.0
redis==5.0.1
python-dotenv==1.0.0
motor==3.3.2
pydantic-settings==2.1.0
fastapi==0.104.1
uvicorn==0.24.0
temporalio==1.4.0
asyncio-mqtt==0.16.1
openai==1.3.0
azure-storage-blob==12.19.0
azure-ai-openai==1.0.0
anthropic==0.7.0
Pillow==10.1.0
opencv-python==4.8.1.78

